Enable sentry for local and production dendrite environments (#1552)
This PR enables Sentry on a new project in our
[org](https://here-not-there.sentry.io/projects/) for Dendrite both
localhost and remote environment where environment is constructed as the
tuple of `node domain, chainId`. Original sentry support was added
upstream by matrix-org in
[this](https://github.com/matrix-org/dendrite/pull/1803) PR.

Test issues:

https://here-not-there.sentry.io/issues/3950993903/?query=is%3Aunresolved&referrer=issue-stream

We should check that issues are published from ECS when this PR is
merged.