cache docker builds (#1775)
The biggest thing standing between us and efficient docker cached builds
is the `RELEASE_VERSION` arg. Since this argument is always changing,
and it's one of the first things used in the Dockerfile, our docker
build cache is always invalidated. This PR addresses that problem via:

1) Remove the -ldflags argument to the go build
2) Move the RELEASE_VERSION arg to the bottom
3) Make dendrite use the environment variable, instead of the build-time
arg
4) Make dockerfile pass the build-time version ARG as an environment
variable into the image

---------

Co-authored-by: Brian Meek <brian@hntlabs.com>